+++ src/api/consumer/index.js
@@ -9,6 +9,7 @@ import { build, buildAll } from './lib/build';
 import importAction from './lib/import';
 import installAction from './lib/install';
 import exportAction from './lib/export';
+import forkAction from './lib/fork';
 import getConsumerComponent from './lib/get-consumer-component';
 import getScopeComponent from './lib/get-scope-component';
 import test from './lib/test';
@@ -39,6 +40,7 @@ export {
   isolate,
   config,
   exportAction,
+  forkAction,
   remove,
   deprecate,
   undeprecate,